requirements.txt
torch==2.4.1
torchvision==0.19.1
transformers==4.37.2
numpy==2.3.2
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
Pillow==11.3.0
ultralytics==8.3.135
bitsandbytes==0.42.0
deepspeed==0.12.6
flash_attn==2.5.5
peft==0.6.2
xformers==0.0.32.post2
pandas==2.3.2
editdistance==0.8.1
tokenizers==0.21.4
pycocoevalcap==1.2
fastapi==0.116.1
gradio==5.43.1
uvicorn==0.35.0
openai==1.101.0
Requests==2.32.5
fire==0.7.1
huggingface_hub==0.32.4
packaging==25.0
shortuuid==1.0.13
tqdm==4.67.1
fitz==0.0.1.dev2
pix2text==1.1.3.2
selenium==4.35.0
webdriver_manager==4.0.2
cog==0.16.5
ray==2.48.0
sglang==0.5.1.post2
